@@ -1,46 +1,54 @@
 {
-    "name": "flist-exported",
-    "version": "1.0.0",
-    "author": "The F-List Team",
-    "description": "F-List Exported",
-    "license": "MIT",
-    "devDependencies": {
-        "@fortawesome/fontawesome-free-webfonts": "^1.0.6",
-        "@types/node": "^10.3.3",
-        "@types/sortablejs": "^1.3.31",
-        "axios": "^0.18.0",
-        "bootstrap": "^4.1.0",
-        "css-loader": "^0.28.11",
-        "date-fns": "^1.28.5",
-        "electron": "^2.0.2",
-        "electron-builder": "^20.8.1",
-        "electron-log": "^2.2.9",
-        "electron-updater": "^2.21.4",
-        "extract-text-webpack-plugin": "4.0.0-beta.0",
-        "file-loader": "^1.1.10",
-        "fork-ts-checker-webpack-plugin": "^0.4.1",
-        "lodash": "^4.16.4",
-        "node-sass": "^4.8.3",
-        "optimize-css-assets-webpack-plugin": "^4.0.0",
-        "qs": "^6.5.1",
-        "raven-js": "^3.24.1",
-        "sass-loader": "^7.0.1",
-        "sortablejs": "^1.6.0",
-        "ts-loader": "^4.2.0",
-        "tslib": "^1.7.1",
-        "tslint": "^5.7.0",
-        "typescript": "^2.8.1",
-        "vue": "^2.5.16",
-        "vue-class-component": "^6.0.0",
-        "vue-loader": "^15.2.4",
-        "vue-property-decorator": "^6.0.0",
-        "vue-template-compiler": "^2.5.16",
-        "webpack": "^4.5.0"
-    },
-    "dependencies": {
-        "@types/lodash": "^4.14.107",
-        "keytar": "^4.2.1",
-        "spellchecker": "^3.4.3",
-        "style-loader": "^0.21.0"
-    }
+  "name": "flist-exported",
+  "version": "1.0.0",
+  "author": "The F-List Team",
+  "description": "F-List Exported",
+  "license": "MIT",
+  "devDependencies": {
+    "@fortawesome/fontawesome-free-webfonts": "^1.0.6",
+    "@types/lodash": "^4.14.116",
+    "@types/node": "^10.5.6",
+    "@types/sortablejs": "^1.3.31",
+    "axios": "^0.18.0",
+    "bootstrap": "^4.1.3",
+    "css-loader": "^1.0.0",
+    "date-fns": "^1.28.5",
+    "electron": "2.0.2",
+    "electron-log": "^2.2.16",
+    "electron-packager": "^12.1.0",
+    "electron-rebuild": "^1.8.2",
+    "extract-text-webpack-plugin": "4.0.0-beta.0",
+    "file-loader": "^1.1.10",
+    "fork-ts-checker-webpack-plugin": "^0.4.4",
+    "lodash": "^4.16.4",
+    "node-sass": "^4.8.3",
+    "optimize-css-assets-webpack-plugin": "^5.0.0",
+    "qs": "^6.5.1",
+    "raven-js": "^3.26.4",
+    "sass-loader": "^7.1.0",
+    "sortablejs": "^1.6.0",
+    "style-loader": "^0.21.0",
+    "ts-loader": "^4.2.0",
+    "tslib": "^1.7.1",
+    "tslint": "^5.7.0",
+    "typescript": "^3.0.1",
+    "vue": "^2.5.17",
+    "vue-class-component": "^6.0.0",
+    "vue-loader": "^15.2.6",
+    "vue-property-decorator": "^7.0.0",
+    "vue-template-compiler": "^2.5.17",
+    "webpack": "^4.16.4"
+  },
+  "dependencies": {
+    "keytar": "^4.2.1",
+    "spellchecker": "^3.4.3"
+  },
+  "optionalDependencies": {
+    "appdmg": "^0.5.2",
+    "electron-squirrel-startup": "^1.0.0",
+    "electron-winstaller": "^2.6.4"
+  },
+  "scripts": {
+    "postinstall": "electron-rebuild -o spellchecker,keytar"
+  }
 }
