@@ -7,19 +7,17 @@ export default class Notifications extends BaseNotifications {
     async notify(conversation: Conversation, title: string, body: string, icon: string, sound: string): Promise<void> {
         if(!this.shouldNotify(conversation)) return;
         try {
-            return super.notify(conversation, title, body, icon, sound);
+            await super.notify(conversation, title, body, icon, sound);
         } catch {
-            (async() => { //tslint:disable-line:no-floating-promises
-                //tslint:disable-next-line:no-require-imports no-submodule-imports
-                await navigator.serviceWorker.register(<string>require('file-loader!./sw.js'));
-                const reg = await navigator.serviceWorker.ready;
-                await reg.showNotification(title, this.getOptions(conversation, body, icon));
-                navigator.serviceWorker.onmessage = (e) => {
-                    const conv = core.conversations.byKey((<{key: string}>e.data).key);
-                    if(conv !== undefined) conv.show();
-                    window.focus();
-                };
-            })();
+            //tslint:disable-next-line:no-require-imports no-submodule-imports
+            await navigator.serviceWorker.register(<string>require('file-loader!./sw.js'));
+            const reg = await navigator.serviceWorker.ready;
+            await reg.showNotification(title, this.getOptions(conversation, body, icon));
+            navigator.serviceWorker.onmessage = (e) => {
+                const conv = core.conversations.byKey((<{key: string}>e.data).key);
+                if(conv !== undefined) conv.show();
+                window.focus();
+            };
         }
     }
 }
